feat: при скролле навбар меняется
